Run a leg of the Applications to Fleets renamathon
Change the term "Application" to "Fleet" or "App" as appropriate. Also meant some revisions to the instructions to "create a new application" each time a new app was to be pushed. Opted to suggest a fleet needs to be created if the reader has not already done so. That way you can re-use the fleet for each new app if you wish.